+++ src/runtime/model/value.rs
@@ -1,8 +1,13 @@
 // Copyright 2024 the Velato Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use vello::kurbo::{self};
-use vello::peniko;
+use vello::{
+    kurbo::{
+        self, ParamCurve, Point,
+        common::{solve_cubic, solve_itp},
+    },
+    peniko,
+};
 
 /// Fixed or animated value.
 #[derive(Clone, Debug)]
@@ -170,16 +175,49 @@ pub trait Tween: Clone {
 }
 
 impl Tween for f64 {
+    /// Tween between two f64 values using a cubic bezier easing curve.
+    ///
+    /// Unfortunately, kurbo doesn't expose a way to get the eased value of t directly, so we have
+    /// to re-implement it here.
+    ///
+    /// t - normalized time 0.0 to 1.0
     fn tween(&self, other: &Self, t: f64, easing: &Easing) -> Self {
-        keyframe::ease(
-            keyframe::functions::BezierCurve::from(
-                keyframe::mint::Vector2::from_slice(&[easing.o.x, easing.o.y]),
-                keyframe::mint::Vector2::from_slice(&[easing.i.x, easing.i.y]),
-            ),
-            *self,
-            *other,
-            t,
-        )
+        // avoid unnecessary calculations
+        if t <= 0.0 {
+            return *self;
+        }
+        if t >= 1.0 {
+            return *other;
+        }
+
+        // Created for [kurbo::ParamCurve::eval]
+        let curve = kurbo::CubicBez::new(
+            Point::new(0.0, 0.0),
+            Point::new(easing.o.x, easing.o.y),
+            Point::new(easing.i.x, easing.i.y),
+            Point::new(1.0, 1.0),
+        );
+
+        // Manual copy-paste of [kurbo::bezpath::cubic_bez_coefs]
+        // in preparation for [kurbo::common::solve_cubic]
+        let (x0, x1, x2, x3) = (curve.p0.x, curve.p1.x, curve.p2.x, curve.p3.x);
+        let c0 = x0;
+        let c1 = 3.0 * (x1 - x0);
+        let c2 = 3.0 * (x2 - 2.0 * x1 + x0);
+        let c3 = x3 - 3.0 * x2 + 3.0 * x1 - x0;
+
+        // Find where along the curve we are at normalized time `t`
+        let roots = kurbo::common::solve_cubic(c0 - t, c1, c2, c3);
+        let u = roots
+            .iter()
+            .copied()
+            .find(|&r| (0.0..=1.0).contains(&r))
+            .expect("a valid curve should always have a root in [0, 1]");
+
+        // Evaluate the curve at that point to get the y value
+        let eased_y = curve.eval(u).y;
+
+        self + (other - self) * eased_y
     }
 }